requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
awsebcli==3.20.3
beautifulsoup4==4.9.3
botocore==1.23.54
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
colorama==0.4.3
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.50.0
django-countries==7.3.2
django-crispy-forms==1.12.0
django-filter==21.1
djangorestframework==3.13.1
future==0.16.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
jmespath==0.10.0
Markdown==3.3.4
markdown2==2.4.0
msgpack==1.0.3
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.2.0
pathspec==0.9.0
Pillow==8.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pypiwin32==223
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.1
pywin32==303
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
semantic-version==2.8.5
service-identity==21.1.0
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.74.0
termcolor==1.1.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.1.9
zipp==3.8.0
zope.interface==5.4.0
